refactor: return view model instead of domain from read persistence
