+++ src/main/java/com/lxp/user/application/port/in/usecase/UserSaveUseCase.java
@@ -1,7 +1,7 @@
-package com.lxp.user.application.in.usecase;
+package com.lxp.user.application.port.in.usecase;
 
 import com.lxp.common.application.port.in.CommandWithResultUseCase;
-import com.lxp.user.application.in.command.UserSaveCommand;
+import com.lxp.user.application.port.in.command.UserSaveCommand;
 
 public interface UserSaveUseCase extends CommandWithResultUseCase<UserSaveCommand, Void> {
 }
