+++ src/main/java/com/lxp/user/infrastructure/persistence/read/adapter/UserReadMapper.java
@@ -1,9 +1,8 @@
 package com.lxp.user.infrastructure.persistence.read.adapter;
 
+import com.lxp.user.application.port.out.query.UserView;
+import com.lxp.user.application.port.out.query.UserWithProfileView;
 import com.lxp.user.domain.common.model.vo.UserId;
-import com.lxp.user.domain.profile.model.entity.UserProfile;
-import com.lxp.user.domain.profile.model.vo.Tags;
-import com.lxp.user.domain.user.model.entity.User;
 import com.lxp.user.domain.user.model.vo.UserEmail;
 import com.lxp.user.domain.user.model.vo.UserName;
 import com.lxp.user.infrastructure.persistence.read.dto.UserDetailDto;
@@ -17,60 +16,41 @@ import java.util.List;
 @Component
 public class UserReadMapper {
 
-    public User toDomain(UserSummaryDto userSummaryDto) {
-        return User.of(
+    public UserView toUserView(UserSummaryDto userSummaryDto) {
+        return new UserView(
             UserId.of(userSummaryDto.id()),
             UserName.of(userSummaryDto.name()),
             UserEmail.of(userSummaryDto.email()),
             userSummaryDto.role().toDomain(),
             userSummaryDto.userStatus().toDomain(),
-            null,
             userSummaryDto.deletedAt()
         );
     }
 
-    public User toDomainWithProfile(UserDetailDto userDetailDto, List<Long> tags) {
-        UserId userId = UserId.of(userDetailDto.id());
-        UserProfile userProfile = toProfile(userDetailDto, userId, tags);
-
-        return User.of(
-            userId,
+    public UserWithProfileView toUserWithProfileView(UserDetailDto userDetailDto, List<Long> tags) {
+        return new UserWithProfileView(
+            UserId.of(userDetailDto.id()),
             UserName.of(userDetailDto.name()),
             UserEmail.of(userDetailDto.email()),
             userDetailDto.role().toDomain(),
             userDetailDto.status().toDomain(),
-            userProfile,
+            userDetailDto.level().toDomain(),
+            tags,
+            userDetailDto.createdAt(),
             userDetailDto.deletedAt()
         );
     }
 
-
-    public UserProfile toProfile(UserDetailDto userDetailDto, UserId userId, List<Long> tags) {
-        return UserProfile.create(
-            userId,
-            userDetailDto.level().toDomain(),
-            new Tags(tags));
-    }
-
-    public User toDomain(UserJpaEntity userEntity, UserProfileJpaEntity profileEntity) {
-        UserId userId = UserId.of(userEntity.getId());
-
-        UserProfile userProfile = null;
-        if (profileEntity != null) {
-            userProfile = UserProfile.create(
-                userId,
-                profileEntity.getLevel().toDomain(),
-                new Tags(profileEntity.getTags())
-            );
-        }
-
-        return User.of(
-            userId,
+    public UserWithProfileView toUserWithProfileView(UserJpaEntity userEntity, UserProfileJpaEntity profileEntity) {
+        return new UserWithProfileView(
+            UserId.of(userEntity.getId()),
             UserName.of(userEntity.getName()),
             UserEmail.of(userEntity.getEmail()),
             userEntity.getRole().toDomain(),
             userEntity.getUserStatus().toDomain(),
-            userProfile,
+            profileEntity.getLevel().toDomain(),
+            profileEntity.getTags(),
+            userEntity.getCreatedAt(),
             userEntity.getDeletedAt()
         );
     }
