+++ src/test/resources/application.properties
@@ -21,3 +21,7 @@ spring.rabbitmq.listener.simple.auto-startup=false
 spring.rabbitmq.host=localhost
 spring.rabbitmq.port=5672
 rabbitmq.queue.auth-events=auth-events.user-registered
+
+# passport
+passport.key.public-key-string=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzPmo5992YpGJ1s2IPxNsF1PamxUi6tN1pC6mmsrtFYBTNbY1G4jRcy7LMQEz0CGvaSvqDW5qCO+JRMhbyisCHjfBGrTwgalY/HDCzoIq+4d8rQV3pK8mQzXh8NsrCorRNBoX8dZZVk22DXXsWUQ+igMgYxo401Hnao7eZ5Twr0NNbbskqWW+jrg5xfPXOeK5bgcGXS8Lzp4uei+lYNokvxP99CIlyN3KkqMiUjJ10rEZH2R/rSLpskkvEQeSQG60J4wuZdereuA6kZJgz1Csg5RDZ0wfaa43vR7UB1GhYfFOwGSXLCypq9XjiPAAv+IiQmwW2UHj0esyp7zyIPsXQwIDAQAB
+
